move "Use another language to call a function"
