--- src/use_another_language_to_call_a_function/.gitignore
@@ -1,1 +1,0 @@
-Cargo.lock
