--- src/use_another_language_to_call_a_function/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "use-another-language-to-call-a-function"
-version = "1.0.0"
-authors = ["Andy Russell, https://github.com/euclio"]
-
-[dependencies]
-libc = "0.2"
-
-[lib]
-crate_type = ["dylib"]
